投稿した人だけの権限設定

+++ app/Http/Controllers/PostController.php
@@ -23,6 +23,9 @@ class PostController extends Controller
    
     public function create()
     {
+    if (Auth::id() !== $post->user_id && !Auth::user()->authority) {
+        return redirect()->back();
+        }
     return view('posts.create');
     }
 
@@ -39,6 +42,9 @@ class PostController extends Controller
     
     public function edit(Post $post)
     {
+       if (Auth::id() !== $post->user_id && !Auth::user()->authority) {
+        return redirect()->back();
+       }
         return view('posts.edit')->with(['post' => $post]);
     }
 
@@ -47,6 +53,9 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {   
+        if (Auth::id() !== $post->user_id && !Auth::user()->authority) {
+        return redirect()->back();
+        }
         $input_post = $request['post'];
         if($request->file('image')){
         $image_url = Cloudinary::upload($request->file('image')->getRealPath())->getSecurePath();
@@ -62,6 +71,9 @@ class PostController extends Controller
          */
      public function delete(Post $post)
     {
+        if (Auth::id() !== $post->user_id && !Auth::user()->authority) {
+            return redirect()->back();
+        }
         $post->delete();
         return redirect('/posts');
     }
